pyproject.toml
[tool.poetry]
name = "tele_photo_changer"
version = "0.1.0"
description = "Telegram photos changer"
authors = ["DzesenKopeykin <kopeykindzzesen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
python-dotenv = "0.19.2"
Pyrogram = "1.2.20"

[tool.poetry.dev-dependencies]
black = "21.12b0"
isort = "5.10.1"
mypy = "0.930"
flake8 = "4.0.1"
pytest = "6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
